controller is now split into two
CarController is use to display car list only now and CarQuoteController is use to display the car quotes
