+++ app/Http/Controllers/CarController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use GuzzleHttp\Client;
 use App\Models\Car;
-use App\Models\Quote;
 
 class CarController extends Controller
 {
@@ -17,7 +16,7 @@ class CarController extends Controller
         $client = new Client();
 
         $url = env('API_URL') . '/phptest/cars';
-        $username = env('API_USERNAME'); // Ensure these are set in your .env file
+        $username = env('API_USERNAME');
         $key = env('API_KEY');
 
         try {
@@ -35,21 +34,21 @@ class CarController extends Controller
                 $cars = $data['cars'];
 
                 // Iterate through each car and store/update in the database
-                foreach ($cars as $car) {
+                foreach ($cars as $carData) {
                     Car::updateOrCreate(
-                        ['license_plate' => $car['licensePlate']], // Unique identifier
+                        ['license_plate' => $carData['licensePlate']], // Unique identifier
                         [
-                            'license_state' => $car['licenseState'],
-                            'vin'           => $car['vin'],
-                            'year'          => $car['year'],
-                            'colour'        => $car['colour'],
-                            'make'          => $car['make'],
-                            'model'         => $car['model'],
+                            'license_state' => $carData['licenseState'],
+                            'vin'           => $carData['vin'],
+                            'year'          => $carData['year'],
+                            'colour'        => $carData['colour'],
+                            'make'          => $carData['make'],
+                            'model'         => $carData['model'],
                         ]
                     );
                 }
 
-                // Retrieve all cars from the database to display
+                // Retrieve all cars from the database to display (pagination can be done in the future if needed)
                 $carList = Car::all();
 
                 return view('index', compact('carList'));
@@ -62,75 +61,4 @@ class CarController extends Controller
         }
     }
 
-    /**
-     * Handle the "Get Quote" action.
-     */
-    public function getQuotes(Request $request)
-    {
-        // Validate incoming data
-        $validated = $request->validate([
-            'licensePlate' => 'required|string|max:255',
-            'licenseState' => 'required|string|max:255',
-        ]);
-
-        $client = new Client();
-
-        $url = env('API_URL') . '/phptest/quotes';
-        $username = env('API_USERNAME');
-        $key = env('API_KEY');
-
-        // Retrieve license details from the form submission
-        $licensePlate = $validated['licensePlate'];
-        $licenseState = $validated['licenseState'];
-
-        try {
-            // Make a POST request to the external API to fetch quotes
-            $response = $client->post($url, [
-                'form_params' => [
-                    'username'      => $username,
-                    'key'           => $key,
-                    'licensePlate'  => $licensePlate,
-                    'licenseState'  => $licenseState,
-                ],
-            ]);
-
-            $data = json_decode($response->getBody(), true);
-
-            if ($data['success'] === 'ok') {
-                $quotesData = $data['quotes'];
-
-                // Find the car in the database
-                $car = Car::where('license_plate', $licensePlate)->first();
-
-                if ($car) {
-                    // Iterate through each quote and store/update in the database
-                    foreach ($quotesData as $quoteData) {
-                        Quote::updateOrCreate(
-                            [
-                                'car_id'           => $car->id,
-                                'repairer'         => $quoteData['repairer'],
-                                'overview_of_work' => $quoteData['overviewOfWork'],
-                            ],
-                            [
-                                'price' => $quoteData['price'],
-                            ]
-                        );
-                    }
-                } else {
-                    return back()->withErrors(['Car not found in the database.']);
-                }
-
-                // Pass both quotes and car details to the view
-                return view('quotes', [
-                    'quotes' => $quotesData,
-                    'car'    => $car,
-                ]);
-            } else {
-                return back()->withErrors(['API call failed.']);
-            }
-        } catch (\Exception $e) {
-            // Handle exceptions such as network issues
-            return back()->withErrors([$e->getMessage()]);
-        }
-    }
 }
