+++ routes/web.php
@@ -2,9 +2,10 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\CarController;
+use App\Http\Controllers\CarQuoteController;
 
 // Home page: Display the car list
 Route::get('/', [CarController::class, 'index']);
 
 // Quotes page: Handle form submissions to get quotes
-Route::post('/quotes', [CarController::class, 'getQuotes']);
+Route::post('/quotes', [CarQuoteController::class, 'getQuotes']);
